problems/src/test/java/_0901_0950/_944_Delete_Columns_to_Make_Sorted/SolutionITest.java
package _0901_0950._944_Delete_Columns_to_Make_Sorted;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 9/26/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final String[] A = {"cba", "daf", "ghi"};
    assertEquals(1, solution.minDeletionSize(A));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final String[] A = {"a", "b"};
    assertEquals(0, solution.minDeletionSize(A));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}